docs(actor-highlight): split js into its own file
